% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_sf}
\alias{df_sf}
\title{A df with a geometry column}
\format{An object of class \code{data.frame} with 1 rows and 3 columns.}
\usage{
df_sf
}
\description{
A df with a geometry column
}
\keyword{datasets}
