% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_sf_tibble}
\alias{df_sf_tibble}
\title{A tibble with a geometry column}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.}
\usage{
df_sf_tibble
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
A tibble with a geometry column
}
\keyword{datasets}
